requirements.txt
dj-config-url==0.1.1
Django==2.2.4
djangorestframework==3.10.2
gunicorn==19.9.0
numpy==1.17.0
pandas==0.25.1
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
